inventory tracking - support for shift-click crafting from the 2x2 crafting field. Works nicely as long as the amount of produced items does not exceed the capacity. If it does, the notching client seem to fill up all slots to stack size and loses the remaining items. We still need to implement this detail

+++ mcp_packet.h
@@ -113,7 +113,7 @@ typedef struct {
 
 typedef struct {
     int16_t item;
-    int8_t  count;
+    int16_t count;  // actually int8_t, but we need to have a larger capacity to deal with crafting
     int16_t damage;
     //nbt_t nbt;
 } slot_t;
